requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.3
elasticsearch==7.10.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
